requirements.dashboard.txt
# LP Dashboard dependencies
streamlit>=1.30.0
httpx>=0.25.0
pandas>=2.0.0
pyyaml>=6.0
plotly>=5.18.0
numpy>=1.24.0
